remove ef core & mssql continued

+++ src/Memodex.WebApp/Common/PagedData.cs
@@ -3,7 +3,7 @@ namespace Memodex.WebApp.Common;
 public class PagedData<TItem>
 {
     public int Page { get; init; }
-    public int ItemCount { get; init; }
+    public long ItemCount { get; init; }
     public int TotalPages { get; init; }
     public IReadOnlyList<TItem> Items { get; init; } = null!;
 }
